+++ qmr.ps1
@@ -1,0 +1,1 @@
+python -m ptb1